+++ changelog.md
@@ -2,7 +2,14 @@
 
 ## [Unreleased]
 
+### Added
+
 * Create a second webpack bundle `factom-struct` containing only factom.js data structures, excluding all online calls related components (FactomCli...). Util for client side applications (browser).
+* Check that transaction amounts are JS safe integers.
+
+### Changed
+
+* Replaced dependency on `bn.js` by `long`.
 
 ## [0.2.6] 2018-07-26